requirements.txt
# Core AWS and AgentCore dependencies
boto3>=1.34.0
botocore>=1.34.0
bedrock-agentcore>=0.1.0
bedrock-agentcore-starter-toolkit>=0.1.0

# Strands framework (Agents as Tools) - Correct package names
strands-agents>=1.19.0
strands-agents-tools>=0.2.17

# MCP (Model Context Protocol)
mcp>=1.0.0

# HTTP client and utilities
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pyyaml>=6.0
retrying>=1.3.4

# OpenSearch (for knowledge base integration)
opensearch-py>=2.7.0

# Development and testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
hypothesis>=6.92.0
black>=23.0.0
ruff>=0.1.0

# Development utilities
ipython>=8.29.0
